+++ lib/pages/transaction/dialogs/currency.dart
@@ -1,9 +1,12 @@
 import 'package:chopper/chopper.dart' show Response;
 import 'package:flutter/material.dart';
+import 'package:logging/logging.dart';
 import 'package:provider/provider.dart';
 import 'package:waterflyiii/auth.dart';
 import 'package:waterflyiii/generated/l10n/app_localizations.dart';
 import 'package:waterflyiii/generated/swagger_fireflyiii_api/firefly_iii.swagger.dart';
+
+final Logger log = Logger("Pages.Transaction.Dialogs.Currency");
 
 class CurrencyDialog extends StatelessWidget {
   const CurrencyDialog({super.key, required this.currentCurrency});
@@ -37,6 +40,8 @@ class CurrencyDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    log.finest(() => "build()");
+
     return SimpleDialog(
       title: Text(S.of(context).transactionDialogCurrencyTitle),
       clipBehavior: .hardEdge,
