+++ lib/pages/transaction/dialogs/delete.dart
